WIP - smooth exit from IT cases during Run goal IT WIP - Refactoring and cleanup
