+++ .gitignore
@@ -12,3 +12,5 @@
 
 # virtual machine crash logs, see http://www.java.com/en/download/help/error_hotspot.xml
 hs_err_pid*
+# this needs to be removed when actual site generation is done
+site
