+++ README.md
@@ -12,7 +12,7 @@ combines the goals of Maven Shade, Maven Exec into one, allowing the users to ha
 | [`vertx-maven-plugin:package`](#) | Package  Vert.X application into a fat jar |
 | [`vertx-maven-plugin:run`](#) | Runs the  Vert.X application forked on to maven JVM (WIP)|
 | [`vertx-maven-plugin:start`](#) | Will start the  Vert.X Application as daemon (WIP)  |
-| [`fvertx-maven-plugin:stop`](#) | Will stop the  Vert.X Application running as daemon (WIP)  |
+| [`vertx-maven-plugin:stop`](#) | Will stop the  Vert.X Application running as daemon (WIP)  |
 
 ### Documentation and Support
