+++ src/main/java/org/workspace7/maven/plugins/AbstractVertxMojo.java
@@ -16,7 +16,6 @@
 
 package org.workspace7.maven.plugins;
 
-import org.apache.commons.io.FilenameUtils;
 import org.apache.maven.artifact.Artifact;
 import org.apache.maven.plugin.AbstractMojo;
 import org.apache.maven.plugin.MojoExecutionException;
@@ -41,22 +40,35 @@ import java.util.Set;
 import java.util.stream.Collectors;
 
 /**
- * @author kameshs
+ * The base Mojo class that will be extended by the other plugin goals
  */
 public abstract class AbstractVertxMojo extends AbstractMojo {
 
     public static final String IO_VERTX_CORE_LAUNCHER = "io.vertx.core.Launcher";
 
     /* ==== Maven deps ==== */
+    /**
+     * The Maven project which will define and confiure the vertx-maven-plugin
+     */
     @Parameter(defaultValue = "${project}", readonly = true)
     protected MavenProject project;
 
+    /**
+     * The project build output directory, defaults to ${project.build.directory} which will be target directory
+     * of the project
+     */
     @Parameter(defaultValue = "${project.build.directory}")
     protected String projectBuildDir;
 
+    /**
+     * The maven artifact resolution session, which will be used to resolve Maven artifacts
+     */
     @Parameter(defaultValue = "${repositorySystemSession}", readonly = true)
     protected DefaultRepositorySystemSession repositorySystemSession;
 
+    /**
+     * The list of remote repositories that will be used to resolve artifacts
+     */
     @Parameter(alias = "remoteRepositories", defaultValue = "${project.remoteArtifactRepositories}", readonly = true)
     protected List<ArtifactRepository> remoteRepositories;
 
@@ -70,12 +82,26 @@ public abstract class AbstractVertxMojo extends AbstractMojo {
     // TODO-ROL: It would be awesome if this would not be required but, if not given,
     // the plugin tries to detect a single verticle. Maybe even decorated with a specific annotation ?
     // (like @MainVerticle ?). Only if no such verticle can be uniquely identified, then throw an exception.
+    /**
+     * The verticle that will be the main entry point on to the vertx application, the same property will be used
+     * as &quot;Main-Verticle;&quot; attribute value on the MANIFEST.MF
+     */
     @Parameter(defaultValue = "${vertx.verticle}", required = true)
     protected String verticle;
 
+    /**
+     * The main launcher class that will be used when launching the Vert.X applications.
+     * It defaults to {@link io.vertx.core.Launcher}
+     */
     @Parameter(defaultValue = "io.vertx.core.Launcher", property = "vertx.launcher")
     protected String launcher;
 
+    /**
+     * this method resolves maven artifact from all configured repositories using the maven coordinates
+     *
+     * @param artifact - the maven coordinates of the artifact
+     * @return {@link Optional} {@link File} pointing to the resolved artifact in local repository
+     */
     protected Optional<File> resolveArtifact(String artifact) {
         ArtifactRequest artifactRequest = new ArtifactRequest();
         artifactRequest.setArtifact(new org.eclipse.aether.artifact.DefaultArtifact(artifact));
@@ -94,6 +120,12 @@ public abstract class AbstractVertxMojo extends AbstractMojo {
         return Optional.empty();
     }
 
+    /**
+     * this method helps in extracting the Artifact paths from the Maven local repository
+     *
+     * @param artifacts - the collection of artifacts which needs to be resolved to local {@link File}
+     * @return A {@link Set} of {@link Optional} file paths
+     */
     protected Set<Optional<File>> extractArtifactPaths(Set<Artifact> artifacts) {
         return artifacts
                 .stream()
@@ -103,6 +135,12 @@ public abstract class AbstractVertxMojo extends AbstractMojo {
                 .collect(Collectors.toSet());
     }
 
+    /**
+     * this method helps in resolving the {@link Artifact} as maven coordinates
+     * coordinates ::= group:artifact:version:[packaging]:[classifier]
+     * @param artifact - the artifact which need to be represented as maven coordinate
+     * @return string representing the maven coordinate
+     */
     protected String asMavenCoordinates(Artifact artifact) {
         // TODO-ROL: Shouldn't there be also the classified included after the groupId (if given ?)
         // Maybe we we should simply reuse DefaultArtifact.toString() (but could be too fragile as it might change
@@ -123,16 +161,13 @@ public abstract class AbstractVertxMojo extends AbstractMojo {
         return artifactCords.toString();
     }
 
-    protected String vertxArtifactName() throws MojoExecutionException {
-        final Artifact artifact = this.project.getArtifact();
-        File artifactFile = getArtifactFile(artifact);
-        String artifactName = artifactFile.getName();
-        StringBuilder vertxArtifactName = new StringBuilder(FilenameUtils.getBaseName(artifactName));
-        vertxArtifactName.append("-fat.jar");
-        return vertxArtifactName.toString();
-
-    }
-
+    /**
+     * This method returns the project's primary artifact file, this method tries to compute the artifact file name
+     *  based on project finalName is configured or not
+     * @param artifact - the project artifact for which the target file will be needed
+     * @return {@link File} representing the project primary artifact
+     * @throws MojoExecutionException
+     */
     protected File getArtifactFile(Artifact artifact) throws MojoExecutionException {
         final String finalName = this.project.getName();
         if (artifact == null) {
